fix segfault in the relay module when initial push fails
When the initial push to a target fails in the relay module we can
segfault since the relay ctx is NULL. THe segfault occurs in the if
which sets up the timer for the push_reconnect method.

In this patch we set the context using the existing create local ctx
method passing in a NULL target as the target is unused in this method.